+++ CHANGES.rst
@@ -8,6 +8,8 @@ Added
 
 - User can chose their favourite display name. :pr:`77`
 - Bumped to authlib 1.2. :pr:`78`
+- Implemented RFC7592 OAuth 2.0 Dynamic Client Registration Management
+  Protocol :pr:`79`
 
 [0.0.14] - 2022-11-29
 =====================